LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/cross-origin-objects/cross-origin-objects-expected.txt

CONSOLE MESSAGE: line 1380: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "http://127.0.0.1:8800". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 1380: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "http://127.0.0.1:8800". Protocols, domains, and ports must match.

PASS Basic sanity-checking 
PASS Only whitelisted properties are accessible cross-origin 
PASS [[GetPrototypeOf]] should return null 
PASS [[SetPrototypeOf]] should throw 
PASS [[IsExtensible]] should return true for cross-origin objects 
PASS [[PreventExtensions]] should throw for cross-origin objects 
PASS [[GetOwnProperty]] - Properties on cross-origin objects should be reported |own| 
FAIL [[GetOwnProperty]] - Property descriptors for cross-origin properties should be set up correctly assert_equals: property descriptor for 0 should be configurable expected true but got false
PASS [[Delete]] Should throw on cross-origin objects 
PASS [[DefineOwnProperty]] Should throw for cross-origin objects 
PASS [[Enumerate]] should return an empty iterator 
FAIL [[OwnPropertyKeys]] should return all properties from cross-origin objects assert_array_equals: Object.getOwnPropertyNames() gives the right answer for cross-origin Window lengths differ, expected 15 got 13
FAIL [[OwnPropertyKeys]] should return the right symbol-named properties for cross-origin objects assert_array_equals: Object.getOwnPropertySymbols() should return the three symbol-named properties that are exposed on a cross-origin Window lengths differ, expected 3 got 0
FAIL [[OwnPropertyKeys]] should place the symbols after the property names after the subframe indices assert_array_equals: Reflect.ownKeys should start with the indices exposed on the cross-origin window. property 0, expected "0" but got "blur"
PASS A and B jointly observe the same identity for cross-origin Window and Location 
PASS Cross-origin functions get local Function.prototype 
PASS Cross-origin Window accessors get local Function.prototype 
PASS Same-origin observers get different functions for cross-origin objects 
PASS Same-origin observers get different accessors for cross-origin Window 
PASS Same-origin observers get different accessors for cross-origin Location 
FAIL {}.toString.call() does the right thing on cross-origin objects Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "http://127.0.0.1:8800". Protocols, domains, and ports must match.
 
